🎨 move ObotManagementClient to types file, single-line import
